[FEATURE] 지원서 양식 수정 #81
- 지원서 양식 confirmYN 필드 추가

+++ module-admin/src/main/java/org/example/domain/application/service/CreateApplicationService.java
@@ -1,6 +1,8 @@
 package org.example.domain.application.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.api_response.exception.GeneralException;
+import org.example.api_response.status.ErrorStatus;
 import org.example.domain.application.controller.request.CopyApplicationRequest;
 import org.example.domain.application.controller.request.UpdateApplicationRequest;
 import org.example.domain.application.Application;
@@ -35,16 +37,19 @@ public class CreateApplicationService {
     );
     createSelectQuestionService.createSelectQuestion(application);
 
-    System.out.println("application.getId() = " + application.getId());
     return CreateApplicationResponse.builder()
       .applicationId(application.getId())
       .build();
   }
 
   /**
-   * 지원서 임시저장
+   * 지원서 저장
    */
   public void updateApplication(Long applicationId, UpdateApplicationRequest request) {
+    if (coreApplicationService.findById(applicationId).getConfirmYN()) {
+      throw new GeneralException(ErrorStatus.BAD_REQUEST, "확정된 지원서는 수정할 수 없습니다.");
+    }
+
     // todo delete query 성능 개선
     applicationRepository.deleteById(applicationId);
 
@@ -52,6 +57,7 @@ public class CreateApplicationService {
       Application.builder()
         .study(coreStudyService.findById(request.studyId()))
         .title(request.title())
+        .confirmYN(request.confirmYN())
         .build()
     );
     createTextQuestionService.updateTextQuestion(application, request.updateTextQuestionList());
@@ -76,7 +82,9 @@ public class CreateApplicationService {
    * 지원서 삭제
    */
   public void deleteApplication(Long applicationId) {
-    coreApplicationService.findById(applicationId);
+    if (coreApplicationService.findById(applicationId).getConfirmYN()) {
+      throw new GeneralException(ErrorStatus.BAD_REQUEST, "확정된 지원서는 삭제할 수 없습니다.");
+    }
     applicationRepository.deleteById(applicationId);
   }
 }
